@@ -1,34 +1,33 @@
 {
   "name": "rica",
-  "version": "1.1.2",
+  "version": "2.0.0",
   "private": true,
   "engines": {
-    "node": ">=16.0.0 <20.0.0"
+    "node": ">=18.0.0"
   },
   "dependencies": {
-    "@blueprintjs/core": "^3.53.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-brands-svg-icons": "^5.15.4",
-    "@fortawesome/free-regular-svg-icons": "^5.15.4",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.15",
-    "@reach/rect": "^0.16.0",
-    "@reach/tabs": "^0.16.4",
-    "chart.js": "3.5.0",
-    "chartjs-plugin-datalabels": "2.0.0",
-    "chartjs-plugin-zoom": "1.1.1",
-    "postcss-cli": "^9.1.0",
-    "react": "^17.0.2",
-    "react-chartjs-2": "3.0.4",
-    "react-dom": "^17.0.2",
-    "react-helmet": "^6.1.0",
-    "react-hotkeys": "^2.0.0",
-    "react-icons": "^4.2.0",
-    "react-papaparse": "^3.17.2",
-    "react-scripts": "^5.0.0",
-    "recharts": "^2.1.2",
-    "styled-components": "^5.3.1",
-    "web-vitals": "^1.1.2"
+    "@fortawesome/fontawesome-svg-core": "^6.5.1",
+    "@fortawesome/free-brands-svg-icons": "^6.5.1",
+    "@fortawesome/free-regular-svg-icons": "^6.5.1",
+    "@fortawesome/free-solid-svg-icons": "^6.5.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@visx/axis": "^3.5.0",
+    "@visx/event": "^3.3.0",
+    "@visx/grid": "^3.12.0",
+    "@visx/group": "^3.3.0",
+    "@visx/responsive": "^3.3.0",
+    "@visx/scale": "^3.5.0",
+    "@visx/shape": "^3.5.0",
+    "@visx/tooltip": "^3.3.0",
+    "@visx/zoom": "^3.3.0",
+    "d3-shape": "^3.2.0",
+    "papaparse": "^5.4.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-helmet-async": "^2.0.4",
+    "react-hotkeys-hook": "^4.4.4",
+    "react-scripts": "^5.0.1",
+    "web-vitals": "^3.5.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -37,8 +36,7 @@
     "test": "react-scripts test",
     "lint": "eslint --ext .js,.jsx src/",
     "lint:fix": "eslint --ext .js,.jsx src/ --fix",
-    "eject": "react-scripts eject",
-    "watch:css": "postcss -w src/styles/tailwind.css -o src/styles/output.css"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
@@ -59,13 +57,11 @@
     ]
   },
   "devDependencies": {
-    "@tailwindcss/jit": "^0.1.10",
-    "autoprefixer": "^10.4.1",
-    "concurrently": "^6.5.1",
-    "gulp": "^3.9.1",
-    "gulp-inline-source": "^3.2.0",
-    "gulp-replace": "^1.1.3",
-    "postcss": "^8.4.5",
-    "tailwindcss": "^2.2.19"
+    "autoprefixer": "^10.4.16",
+    "gulp": "^4.0.2",
+    "gulp-inline-source": "^4.0.0",
+    "gulp-replace": "^1.1.4",
+    "postcss": "^8.4.32",
+    "tailwindcss": "^3.4.0"
   }
 }
