@@ -1,38 +1,43 @@
-import React from "react";
-import CarpetOption from "./CarpetOption";
+import React, { useState, useCallback } from "react";
 
-class Carpets extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      carpetPlot: this.props.images[0]["img"],
-    };
-  }
+function Carpets({ images }) {
+  const [carpetPlot, setCarpetPlot] = useState(images?.[0]?.img || "");
 
-  onChange = (e) => {
-    this.setState({ carpetPlot: e.target.value });
-  };
+  const handleChange = useCallback((e) => {
+    setCarpetPlot(e.target.value);
+  }, []);
 
-  render() {
-    const images = this.props.images;
-
+  if (!images?.length) {
     return (
-      <center>
-        <select className="dd-menu" onChange={this.onChange}>
-          {images.map((carpet) => (
-            <CarpetOption
-              key={carpet.name}
-              name={carpet.name}
-              img={carpet.img}
-            />
-          ))}
-        </select>
-        <div className="carpet-plots-image">
-          <img id="imgCarpetPlot" alt="" src={this.state.carpetPlot} />
-        </div>
-      </center>
+      <div className="flex items-center justify-center h-64">
+        <p className="text-gray-500">No carpet plots available</p>
+      </div>
     );
   }
+
+  return (
+    <center>
+      <select
+        className="dd-menu text-gray-700 text-base p-2.5 rounded-md border border-gray-300 mt-4 ml-4 focus:ring-sky-500 focus:border-sky-500"
+        onChange={handleChange}
+        value={carpetPlot}
+      >
+        {images.map((carpet) => (
+          <option key={carpet.name} value={carpet.img}>
+            {carpet.name}
+          </option>
+        ))}
+      </select>
+      <div className="carpet-plots-image mt-4">
+        <img
+          id="imgCarpetPlot"
+          alt="Carpet plot"
+          src={carpetPlot}
+          className="max-w-full transition-opacity duration-300"
+        />
+      </div>
+    </center>
+  );
 }
 
 export default Carpets;
