@@ -1,5 +1,4 @@
-import React, { Component } from "react";
-import { Switch, SwitchLabel, SwitchSelection } from "./ToggleStyles.js";
+import React, { useCallback } from "react";
 
 const titleCase = (str) =>
   str
@@ -7,63 +6,41 @@ const titleCase = (str) =>
     .map((w) => w[0].toUpperCase() + w.slice(1))
     .join(" ");
 
-const ClickableLabel = ({ title, onChange, id }) => (
-  <SwitchLabel onClick={() => onChange(title)} className={id}>
-    {titleCase(title)}
-  </SwitchLabel>
-);
-
-const ConcealedRadio = ({ value, selected }) => (
-  <input
-    className="hidden"
-    type="radio"
-    name="switch"
-    checked={selected === value}
-    readOnly
-  />
-);
-
-class ToggleSwitch extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      selected: this.props.selected,
-    };
-  }
-
-  componentDidUpdate() {
-    if (this.state.selected !== this.props.selected) {
-      this.setState({ selected: this.props.selected });
-    }
-  }
-
-  selectionStyle = () => {
+function ToggleSwitch({ values, selected, colors, handleNewSelection }) {
+  const selectionStyle = useCallback(() => {
+    const index = values.indexOf(selected);
     return {
-      left: `${(this.props.values.indexOf(this.state.selected) / 3) * 100}%`,
-      background: `${
-        this.props.colors[this.props.values.indexOf(this.state.selected)]
-      }`,
+      left: `${index * 90}px`,
+      background: colors[index],
     };
-  };
+  }, [values, selected, colors]);
 
-  render() {
-    return (
-      <div className="relative h-8 -mt-1 font-semibold bg-gray-200 rounded-md w-fit">
-        {this.props.values.map((val) => {
-          return (
-            <span>
-              <ConcealedRadio value={val} selected={this.state.selected} />
-              <ClickableLabel
-                title={val}
-                onChange={this.props.handleNewSelection}
-              />
-            </span>
-          );
-        })}
-        <SwitchSelection style={this.selectionStyle()} />
-      </div>
-    );
-  }
+  return (
+    <div className="relative h-8 font-semibold bg-gray-200 rounded-md flex">
+      {values.map((val) => (
+        <span key={val}>
+          <input
+            className="hidden"
+            type="radio"
+            name="switch"
+            checked={selected === val}
+            readOnly
+          />
+          <label
+            onClick={() => handleNewSelection(val)}
+            className="relative z-10 h-8 flex items-center justify-center transition-colors duration-200"
+            style={{ width: 90, color: selected === val ? "#fff" : "rgba(0,0,0,0.6)", cursor: "pointer" }}
+          >
+            {titleCase(val)}
+          </label>
+        </span>
+      ))}
+      <span
+        className="absolute top-0 z-0 block h-8 rounded-md transition-all duration-200"
+        style={{ ...selectionStyle(), width: 90 }}
+      />
+    </div>
+  );
 }
 
 export default ToggleSwitch;
